Maintain Timeline Fragment - commit 2

+++ app/src/main/java/com/thinkdiffai/futurelove/view/adapter/HistoryAdapter.java
@@ -30,7 +30,7 @@ public class HistoryAdapter extends RecyclerView.Adapter<HistoryAdapter.HistoryV
 
     public interface IOnClickItemListener {
 
-        void onClickItem(long id);
+        void onClickItem(int id);
         void onClickDelete(int orderNumber_id);
 
     }
